updatedSharekhan/requirements.txt
# Core dependencies for SharekhanAPI
requests>=2.28.0
six>=1.16.0
pycryptodome>=3.20.0
cryptography>=39.0.0
websocket-client>=1.8.0
shareconnect
pyotp>=2.6.0

# Flask Web Application Dependencies
Flask>=2.3.3
Flask-CORS>=4.0.0
PyJWT>=1.7.0,<3.0.0
pandas>=2.1.3

# Optional for automated browser auth
selenium>=4.15.2

